Refactor some test to not hog memory
